libkrun: Pass PREFIX to build, fixes library ID
During the build, the libkrun Makefile uses install_name_tool to set the
ID of the installed library. This call uses $PREFIX, but it was not set
and therefore defaulted to /usr/local.

Binaries built against this library will fail to locate libkrun.1.dylib,
because the library ID is copied into the dependent binary during
linking and this path is where the library is then expected at runtime.